Split settings.rs and app/llm.rs into focused submodules
The Gemini-OAuth + provider work grew both files past 1000 lines.
Carve each one by concern so the parent stays a thin entry point:

- crates/mogen-studio/src/settings/{provider_slot,image_provider,paths}.rs
  hold the slot enum + OAuth model defaults, the image-provider
  preference, and the settings.json path resolver respectively. The
  parent settings.rs already references these via `mod` declarations
  (added in 1715208 alongside the keyring removal) but the submodule
  files themselves had not been tracked — without this commit
  settings.rs does not compile from a fresh clone.

- crates/mogen-studio/src/app/llm/{credentials,enhance,poll,spawn,
  textures}.rs hold the credential resolver, prompt-enhance side
  channel, worker-channel polling + outcome application, the text-LLM
  spawn path, and the texture-pipeline starters. app/llm.rs keeps
  only the public starters (generate/modify/repair/animate), retry,
  cancel, and in-flight counters — mirroring the app/util.rs +
  app/util/ convention already established in the studio.

No behaviour change. `cargo check --workspace` is clean and the
mogen-studio test suite still passes (184/184).
